requirements.txt
flask
requests
python-dotenv
cryptography
werkzeug